move updateChampionData to match_presenter
